Move Pymunk LunarLander into environment package
