+++ gymnasium/envs/pymunk/__init__.py
@@ -1,0 +1,5 @@
+"""Pymunk-based environments."""
+
+from gymnasium.envs.pymunk.lunar_lander import LunarLander
+
+__all__ = ["LunarLander"]
